+++ lab1/src/entities/Student.java
@@ -1,17 +1,17 @@
 package entities;
 
 public class Student {
-    private int       _id;
+    private long      _id;
     private Subject[] _attendingSubjects;
     private Mark[]    _marks;
 
-    public Student(int _id, Subject[] _attendingSubjects, Mark[] _marks) {
+    public Student(long _id, Subject[] _attendingSubjects, Mark[] _marks) {
         this._id                = _id;
         this._attendingSubjects = _attendingSubjects;
         this._marks             = _marks;
     }
 
-    public int get_id() {
+    public long get_id() {
         return this._id;
     }
 
@@ -23,7 +23,7 @@ public class Student {
         return this._marks;
     }
 
-    public void set_id(int _id) {
+    public void set_id(long _id) {
         this._id = _id;
     }
