+++ lab1/src/entities/University.java
@@ -1,15 +1,15 @@
 package entities;
 
 public class University {
-    private int       _universityId;
+    private long      _universityId;
     private Faculty[] _faculties;
 
-    public University(int _universityId, Faculty[] _faculties) {
+    public University(long _universityId, Faculty[] _faculties) {
         this._universityId = _universityId;
         this._faculties = _faculties;
     }
 
-    public int get_universityId() {
+    public long get_universityId() {
         return this._universityId;
     }
 
@@ -17,7 +17,7 @@ public class University {
         return this._faculties;
     }
 
-    public void set_universityId(int _universityId) {
+    public void set_universityId(long _universityId) {
         this._universityId = _universityId;
     }
